+++ CHANGELOG.md
@@ -1,3 +1,15 @@
-0.0.1
+0.0.3
 -------------------
+* Fixed `open_pool` issues
+* Added tests with code coverage
+
+0.0.2 [YANKED]
+-------------------
+Yanked due to `open_pool` issues
+* Fixed dependencies
+
+
+0.0.1 [YANKED]
+-------------------
+Yanked due to dependencies problem
 * Initial release